move functions to avoid circular import issues
